refactor(compose): the derivation's execution half moves to its own file
The mask exclusion pushed derivation.go past the 500-line cap; the fetch
(rows + aggregate recompute + shared WHERE) reads as one concept apart
from the compile half. The integration test's one any-typed coercion
seam carries its waiver.
